+++ .gitignore
@@ -9,7 +9,6 @@
 *.user
 *.userosscache
 *.sln.docstates
-Logs.json
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
