+++ CryptoSoft/CryptoSoft/Program.cs
@@ -43,7 +43,8 @@ namespace CryptoSoft
                             // Writing of the encrypted content in a new file
                             File.WriteAllBytes(encryptedFileName, encryptedBytes);
 
-                            fileInfo.Delete();
+                            //temporary disable
+                            //fileInfo.Delete();
                         }
                         // Decryption
                         else if (args[0] == "-d")
