+++ EasySaveWPF/EasySaveClasses/ViewModelNS/Logs.json
@@ -31,4 +31,61 @@
     "DirSize": 0.0,
     "DirTransferTime": 25.0
   },
+  {
+    "Timestamp": "19/02/2024 19:44:25",
+    "Name": null,
+    "sourcePath": null,
+    "targetPath": null,
+    "DirSize": 0.0,
+    "DirTransferTime": 25.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:40:33",
+    "Name": "w-02-19-2024-7-32-38",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\w",
+    "targetPath": "C:\\Users\\33782\\Downloads\\p\\q\\w-02-19-2024-7-32-38",
+    "DirSize": 0.0,
+    "DirTransferTime": 0.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:40:40",
+    "Name": "w-02-19-2024-7-32-38",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\w",
+    "targetPath": "C:\\Users\\33782\\Downloads\\p\\q\\w-02-19-2024-7-32-38",
+    "DirSize": 0.0,
+    "DirTransferTime": 0.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:43:16",
+    "Name": "cacatest-02-19-2024-8-42-56",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\cacatest",
+    "targetPath": "C:\\Users\\33782\\Downloads\\pip\\cacatest-02-19-2024-8-42-56",
+    "DirSize": 1498.0,
+    "DirTransferTime": 1498.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:44:31",
+    "Name": "cacatest-02-19-2024-8-42-56",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\cacatest",
+    "targetPath": "C:\\Users\\33782\\Downloads\\pip\\cacatest-02-19-2024-8-42-56",
+    "DirSize": 202241.0,
+    "DirTransferTime": 1498.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:49:21",
+    "Name": "cacatest-02-19-2024-8-49-18",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\cacatest",
+    "targetPath": "C:\\Users\\33782\\Downloads\\pip\\cacatest-02-19-2024-8-49-18",
+    "DirSize": 202241.0,
+    "DirTransferTime": 202241.0
+  },
+  {
+    "Timestamp": "19/02/2024 20:50:37",
+    "Name": "cacatest-02-19-2024-8-49-18",
+    "sourcePath": "C:\\Users\\33782\\Downloads\\p\\cacatest",
+    "targetPath": "C:\\Users\\33782\\Downloads\\pip\\cacatest-02-19-2024-8-49-18",
+    "DirSize": 202241.0,
+    "DirTransferTime": 0.0,
+    "CryptingTime": 6257
+  }
 ]
