+++ EasySaveWPF/EasySaveWPF/ViewNS/Home.xaml.cs
@@ -24,12 +24,14 @@ namespace EasySaveWPF.ViewNS
     {
         MainWindow _mainWindow;
         MainViewModel _mainViewModel;
-        public Home(MainWindow mainWindow)
+        PlayBreak _playBreak;
+        public Home(MainWindow mainWindow, PlayBreak playBreak)
         {
             InitializeComponent();
             _mainWindow = mainWindow;
             _mainViewModel = _mainWindow.mainViewModel;
             DataContext = _mainViewModel;
+            _playBreak = playBreak;
         }
 
         private void ButtonDelete_Click(object sender, RoutedEventArgs e)
