+++ EasySaveWPF/EasySaveWPF/ViewNS/MainWindow.xaml.cs
@@ -25,24 +25,33 @@ namespace EasySaveWPF.ViewNS
     public partial class MainWindow : Window
     {
         public MainViewModel mainViewModel;
-        Home home;
-        Create create;
-        PlayBreak playBreak;
-        Execution execution;
-        Settings settings;
+        PlayBreak _playBreak;
         public MainWindow()
         {
             InitializeComponent();
             LoadLanguage_En();
-            mainViewModel = new MainViewModel();
-
-            home = new Home(this);
-            create =  new Create(this);
-            Frame.Navigate(home);
             Dispatcher mainDispatcher = Dispatcher.CurrentDispatcher;
-            execution = new Execution(this);
-            settings = new Settings(this);
-
+            mainViewModel = new MainViewModel();
+            PlayBreak playBreak = new PlayBreak(this);
+            Home.Navigate(new Home(this, playBreak));
+            Create.Navigate(new Create(this));
+            PlayBreak.Navigate(new Execution(this));
+        }
+        private void ComboBox_LanguageChanged(object sender, SelectionChangedEventArgs e)
+        {
+            ComboBox comboBox = sender as ComboBox;
+            ComboBoxItem selectedItem = comboBox.SelectedItem as ComboBoxItem;
+            if (selectedItem != null)
+            {
+                if (selectedItem.Content.ToString() == "Francais")
+                {
+                    LoadLanguage_Fr();
+                }
+                if (selectedItem.Content.ToString() == "English")
+                {
+                    LoadLanguage_En();
+                }
+            }
         }
         private void LoadLanguage(string relativePath)
         {
@@ -56,12 +65,12 @@ namespace EasySaveWPF.ViewNS
             }
         }
 
-        public void LoadLanguage_Fr()
+        private void LoadLanguage_Fr()
         {
             LoadLanguage("../../../Properties/French.xaml");
         }
 
-        public void LoadLanguage_En()
+        private void LoadLanguage_En()
         {
             LoadLanguage("../../../Properties/English.xaml");
         }
@@ -75,24 +84,5 @@ namespace EasySaveWPF.ViewNS
         {
             LoadLanguage_Fr();
         }
-
-        private void RadioButtonHome_Click(object sender, RoutedEventArgs e)
-        {
-            Frame.Navigate(home);
-        }
-        private void RadioButtonCreate_Click(object sender, RoutedEventArgs e)
-        {
-            Frame.Navigate(create);
-        }
-
-        private void RadioButtonSetting_Click(object sender, RoutedEventArgs e)
-        {
-            Frame.Navigate(settings);
-        }
-        private void RadioButtonExecution_Click(object sender, RoutedEventArgs e)
-        {
-            Frame.Navigate(execution);
-        }
-        
     }
 }
