+++ EasySaveWPF/EasySaveWPF/ViewNS/PlayBreak.xaml
@@ -6,7 +6,6 @@
         xmlns:local="clr-namespace:EasySaveWPF.ViewNS"
         mc:Ignorable="d"
         Title="Loading" Height="150" Width="400">
-    
     <Grid>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="*" />
