--- EasySaveWPF/EasySaveWPF/ViewNS/Theme/ModernScrollBar.xaml
@@ -1,23 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-
-    <Style x:Key="ModernScrollBar" TargetType="ScrollBar">
-        <Setter Property="Background" Value="#FFFFFF"/>
-        <Setter Property="Foreground" Value="#37474F"/>
-        <Setter Property="Width" Value="10"/>
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="ScrollBar">
-                    <Grid>
-                        <Border Background="{TemplateBinding Background}" CornerRadius="5"/>
-                        <Track Name="PART_Track">
-                            <Track.Thumb>
-                                <Thumb Background="{TemplateBinding Background}" BorderBrush="{TemplateBinding Foreground}" BorderThickness="1" Height="30"/>
-                            </Track.Thumb>
-                        </Track>
-                    </Grid>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-    </Style>
-</ResourceDictionary>
